Introduce :smartcase named argument
In principle for all string related functions that also have :ignorecase.
In this commit, only "starts-with" has been equipped with it.  Basically
it operates as :ignorecase if the needle does **not** have any uppercase
characters.  Otherwise it is ignored.

    say "Foo".starts-with("foo",:smartcase);  # True
    say "Foo".starts-with("Foo",:smartcase);  # True
    say "foo".starts-with("Foo",:smartcase);  # False

If accepted, this PR will be expanded to cover all appropriate methods.