Make feed error message more comprehensible
As discussed in https://github.com/rakudo/rakudo/issues/5061 :

    % raku -e 'my @array[3,3] <== [0..8]'
    ===SORRY!=== Error while compiling -e
    Cannot feed into shaped arrays, as one cannot '.append' to them.

The check for this is a little fragile, as it assumes that a variable
that starts with '@' and which cannot .append, is a shaped array.  This
is not necessarily always the case, although I cannot come up with a
way to do this.  Until someone does, I think this solution is fine
until we can probably do this better in the RakuAST branch.

Also split two other error messages across two lines, as they were
definitely too long for a single line.