Introduce List.over-iterator
This is a counterpart to from-iterator in the same sense that a clone is
a counterpart to new. If a subtype adds information to List on an
instance (e.g. Array and its container descriptor, shape), this is lost
with from-iterator, but is preserved in with-iterator, whose job is to
clone while scrubbing elements and swapping that in. While from-iterator
could theoretically take care of this instead, from-iterator is depended
on by Perl6::Actions, implying any addition to it would be costly, and
existing code may depend on extra container information getting erased.

This is an incomplete feature without an Array.from-iterator.