Keep version file when cleaning an installation
This commit changes the CompUnit::Repository::Staging.remove-artifacts
method to not remove a repo's `version` file.  This file is required to
allow subsequent installations to the same directory; its absence caused
`install-dist.raku` to fail in odd ways (see issue #4907).  From
searching the codebase, I believe that `install-dist.raku` is the only
user of the `remove-artifacts` method, so this change shouldn't have
much impact.