Fix List.STORE usage of Reifier.reify-until-lazy
List.is-lazy also depends on this method, but instead of comparing
against IterationEnd, it makes another statement checking if the list is
fully reified. It does this because reify-until-lazy has a more uniform
type than Iterator.push-until-lazy in always returning the total reified
elements. This never actually clearing the reifier punishes eagerly
reifiable lists for no reason.