Introduce List.imbue, a less eager reification
This is equivalent to the eager method in the case of an eagerly
reifiable list; should a reification ever become lazy however, it will
stop at that point, in which case imbue won't converge with eager until
the tail of the list is needed.

This is added particularly to address a discrepancy between
make-iterator and from-iterator/over-iterator in that make-iterator will
cheat and reify whatever portion of the list it can to make it more
usable from custom STORE implementations. As far as the ecosystem goes,
Data::Record could use this in combination with over-iterator to fix its
reification-related errors.

I see this as being in the same vein of method as cache, eager, and
lazy. There would be more work needed elsewhere to make this uniform
with those (Any/Sequence methods, syntax), but it's arguably an edge
case specific to List, so leave that alone for now.