Array.from-iterator becomes List.from-iterator
We have a definition for Array.make-iterator now. List.from-iterator is
defined in make-iterator's terms. Because Array's $!descriptor defaults
to the default_cont_spec HLL symbol already, but because it also pushes
from its iterator until lazy, Array.from-iterator should behave like a
make-iterator + imbue as implemented. This commit eschews the imbue
portion of Array.from-iterator in eliminating it altogether in order to
give the *-iterator constructors more uniform logic.