List's StructuredIterator is overkill; inline it away
With regards to StructuredBuffer, it's a little paranoid about pulling
elements from its source. Except from within its lexical &assign, it
doesn't need to check the vision because that can never have elements by
the time it reaches any branch that wants to bind a value. We'll mark
such an event with a Shift value so its up to the caller what happens.

While this complicates StructuredIterator.pull-one as implemented, this
makes it possible to shove the workload into a push-until-lazy method,
reducing pull-one to a pull-one on its source, but this point, what's
the point of having the StructuredIterator if it's just going to wrap
and do nothing else eventually?

We'll optimize over make-iterator in STORE's case by introducing two
more methods: make-iterable and make-itemized. In the case of the item,
we already know exactly how an assignment will play out, so we don't
need to bother with laziness; in the case of the iterable's iteration,
it can be preserved as provided while dealing with the destructuring
eagerly. In either case, we get some composable components with which to
keep STORE small.