Support laziness in List.new
List.new lacks support for a lazy Slip as an argument, which can be
produced via coercion or the from-iterator family of constructors. While
we could make this a straight CREATE and STORE chain, this comes with a
punishing overhead for eager slips compared to the current approach.
Luckily, our raw slurpy argument happens to be a List that we can imbue
already for List's sake; subtypes can benefit from a little consistency
with regards to STORE in new compared to Array etc. instead.