Upgrade Array.from-list to List.from-list, a reifying constructor
Making from-iterator a uniformly lazy constructor has its consequences.
This is expected to be capable of reifying any eagerly knowable segment
of a list as an equivalent to Array.from-list. Because we don't do that
anymore, but we can also benefit from of expanding its type to allow it
to serve as the reifying constructor for iterables, we define this on
List instead of Array with its make-iterable method. This is typed as
Iterable:D for arrays' sake, who suffer a slower reification under Mu
without any means to imbue.