Upgrade List.imbue to Iterable.imbue, a more lazily reifying coercion
This method requires that the base has both laziness and a notion of
there being an eager head to this in order to be of any use, so a true
Iterable is needed in place of an Any that can only behave like an eager
singleton. This will allow Array.from-list to perform more consistently.