Reduce List's StructuredBuffer to a routine, fixing off-by-one deconts
StructuredBuffer would attempt to predict that there would be exactly as
many deconts as there are elements on the LHS. This means that for each
list on the LHS, at least its head would risk winding up with a decont
that shouldn't occur. Because its destructuring algorithm depends so
heavily on how its prior project behaves, we'll scrap it altogether.

The lexical &destructure combines StructuredBuffer's project and
destructure methods. It still produces a vision buffer, though it will
process this at the same time as Whatever and STORE calls. A second pass
over indices to be kept after the destructuring performs the remaining
assignments. This should behave comparably to its prior implementation,
but somehow, this hinges on its integer buffering. Other approaches to
caching assignments backed by object buffers don't benchmark so hot.