Give Array::Shaped its own ring-shaped reification-target
The new make family of constructors for List are apt to depend on the
reification-target of an Array, but this is incompatible with shaped
arrays' reification because of its MultiDimArray REPR. We introduce a
RingShapedReificationTarget to manage counting of indices to bind to
incrementally in lieu of pushing. Despite the shape, this is unbounded
in that it repeats from the zeroth position after pushing to the end of
a shaped array because this skips extra bounds checks that would be
necessary otherwise, and are dealt with elsewhere anyway.