Better wrangle containers of RHS elements in native/shaped array STORE
We can take a more direct path to a copy of an element in STORE given a
native or shaped array: clobber any existing elements by always binding
a new container, assigning elements (or Nil to mark their absence)
directly. We can't recycle elements when we need to rebind the entire
reification and Nil will become the default value for the container
anyhow, preserving prior behaviour.