[JVM] Fix defaults for typed optional @/% params
This intends to port 7148da1c61 to the JVM backend.

Please note the workaround for handling @deprecation during compilation
of settings. I'm not really sure why this is needed, but I think that
the problem arises from "sub DYNAMIC" having @deprecation as an
optional parameter in src/core.c/stubs.pm6. For some reasons the check
for gcx.Positional fails in that situation. My wild and vague guess
would be that the check happens too early and some important pieces
are still missing.