Restructure internal shaped array leaves
There are a handful of closely-related Rakudo::Iterator::ShapeLeaf
subtypes floating around in Array::Shaped and array. These perform the
one-to-one assignments involved in their STORE candidates on either
type, for either type.

These share a theme in name in the sense that they all have CPY tacked
to the end of it. We can rename all of these to Copy with the variations
on their types being expressed similarly to trait_mod to help clarify.

That's all the work that's needed for array with its two shape leaves
for now, but Array::Shaped is a little more involved. It has three, but
two more are low-hanging fruit, so their constructors will get pretty
obnoxious. These differ on the type of the source array, so we express
their related construction patterns in roles. This reduces the
differences between the types to a mixin and one method handling their
one problem bind opcode.

array will probably need a similar structure to Array some day, but is
also the reason we don't add the two missing shape leaves for
Array::Shaped at the same time, because it takes some otherwise
unrelated build script muckery.