RakuAST: ensure that repossessions during BEGIN time are recorded
The VM only records repossessions if an SC is on the stack of current
compilations as doing so costs a bit of performance and is only required
if we're actually producing a precomp file. We pushed the SC too late,
missing BEGIN time effects like module loading which can lead to
modified and repossessed Stashes. So follow the old frontend in pushing
the SC at the very start of compilation.