Introduce RAKUDO_GC_EVERY environment variable
This affects whether a GC run will be forced for every N ThreadPoolScheduler
supervisor runs (which is usually at about 100/second).  A value of 1 thus
would force GC runs 100x/second, whereas a value of 100 would force a GC
run every second.

Note that for this to work, the supervisor must actually be running.
The simplest way to force that is, is doing a `start { }`.

The default for RAKUDO_GC_EVERY is effectively to never do an additional
GC run.

One could think about starting the supervisor automatically if
RAKUDO_GC_EVERY is explicitely specified.