Introduce Rakudo::Iterator::FrameIterable for shaped array STORE
Array::Shaped.STORE's Iterable:D candidate currently depends on
StoreIterable, a mixin of Rakudo::Iterator::ShapeBranch. This role
keeps track of offsets in a target list and delegates a pull-one to a
process method, sanitizing offsets it tracks after changes expected of
this, tearing down from its done method if marked for completion. This
is well-suited to algorithms that iterate a shape in an arbitrary order.

StoreIterable is not one of these algorithms. Its implementation is
aware that such validation is more often than not unnecessary in its
case and batches each element in a leaf of the shape in one process call
in an attempt to skirt this. We can take a more predictable path through
reification that lacks the need for this altogether instead by making
a reification target keep track of offsets for us. In order to pull this
off, we need to be less paranoid about how we iterate the elements on
the RHS. Provided that an iteration does not change the structure of
itself in-place by laziness, a push-exactly and pull-one can eliminate a
bunch of countres required for ShapeBranch's stateful pull-one approach.
This can make for a modest performance improvement in assignments of
shapeless iterables to shaped arrays.