Streamline flow of Array construction
Array.new generally chains a set-shape and STORE call. set-shape is an
implementation detail that's more complicated than an nqp::create;
similarly to ^parameterize, it's in charge of performing any mixin a
shaped array requires, but like bless, it creates a new array and
prepares its state given the type produced. We can split this into two
separate operations to the benefit of any constructor, not just new.
These are designed to be low-level, but safe enough to be public API.

The dim constructor generally takes the place of set-shape as a means of
embedding a shape into a fresh array. It should be possible to preserve
any existing dependent naughtiness in the ecosystem while shifting
set-shape's work over to dim by giving this a multi with the laxest type
available. This is now depended on by new in the presence of a shape.

Unlike set-shape, the rub constructor tends to be delegated to by dim
when it needs to produce an empty array while copying over the provided
shape. This was, and still is the Whatever candidate of set-shape, but
splitting this from dim makes for easier dispatch in other parts of this
hot method. This is now depended on by new in the absence of a shape.

rub will also copy over any knowable descriptor from the invocant. This
makes it possible to simplify shaped arrays' constructors so as to keep
preserving any existing type, implying typed arrays can also depend on
this to set up their default descriptor on an instantiation of a type
object, without having to override all of new. There are a handful of
places this can help in, but we'll just focus on the new method for now.