Array[Mu] makes an untyped Array
The default of/keyof on a Hash makes an untyped Hash. Array is
inconsistent with Hash in the sense that Array[Mu] does *not* make a
bare Array, but literally types elements as Mu as a typed array. Make
this untyped so an Array can always be equivalent to a parameterization
with its of; this makes a nicer flow of construction in new possible.