Make List.new a proto w/ a candidate for the empty list
This make the case of an empty new call on list more performant, e.g.
when defining an @-sigilled variable to be a List, regardless of any
STORE-backed assignment that follows separately. We don't go full Array
and add an equivalent single element candidate at the same time because
Roast depends on List.new subsuming those as individual elements, in
which case we lose the ability to depend on STORE optimally, which is
the main reason such a multi would be nice to begin with.