HLLize RI.SHAPED-ARRAY-STORAGE somewhat
This method is responsible for creating the initial reification of the
eager shaped array, where it cannot be Array's nqp::null. It validates
that elements of a shape are compatible with nqp::setdimensions by
chaining typechecks on each element. It coerces elements to Int, but
only when there's one element to consider (never, as it's used).
Interestingly, it parameterizes a knowhow singleton in order to cache
the type of the reification that it needs to calculate at the same time.

We'll be reapplying the singleton knowhow parameterization in the
arrays' dim methods. This can be written more clearly with the "knowhow"
keyword. Likewise, the typechecks can be handled efficiently enough as a
proto, which have been touched on by new-disp. Given that we don't make
any, drop Int coercions to deal in List:D only.