Fix elems handling of list reifications in array dimensioning
Originally, set-shape allowed for int overflows of shape elems. This
should be an int because it's forwarded to an nqp::atpos later, but it
needs to check for the upper bounds of allowable dimensions before it
attempts to use those as an index in the hardcoded Shaped role range.
The splitting of set-shape into rub and dim does not fix this, and
compounds the issue by introducing a regression in the untested event of
a 0D shape. Such a List can lack a reification, so ensure one exists.
In either case, attempt to describe the VM dimension range in the error.