Get more smartmatch cases done right
As covered by spec S03-smartmatch/00-sanity.t. What fails for now is
non-matching regex where `Nil` is expected but it gets boolified. In
order for this to be sorted out `RakuAST::Infix` must know its LHS child
and there is no legal way to have this information for now.