Cache template instances to clone natively in a shaped Array.dim
As with RI.SHAPED-ARRAY-STORAGE, we parameterize a singleton knowhow for
its thread-safe object cache. Similarly, we're stuck doing a bizarre
nqp::setelems on the list of type arguments in order to forward them in
a reliably cacheable way, but instead of a type, we cache a rub. If
compatible with nqp::create, the array should be compatible with
nqp::clone in its initial state as well, skipping successive rub calls.