Handle list coercions in array dimensioning more like before
Coercions at the type level tank the performance of the dim method that
exists on arrays. This is an attempt to avert a scenario where a Mu
instance lacking a List method would make for an LTA error message, but
this convenience should not punish the more common scenario of an
argument threading to Any then List successfully, e.g. a Junction that
breaks dispatch. Trust a Mu dependent to understand errors from here.

&samewith, for some reason, seemd to forward the mixin base instead of a
mixin in the case of a typed array. Re-dispatch altogether.