Refer to "shaped" Mu:U anonymously in Array.dim
The default $!descriptor in an Array is named "element", giving
new on an Array:U a strange, syntactically invalid 'my T element' error.
This is not a revert because @foo is probably also not what the user
wrote; while not ideal, let the user glean the name based on the type.