Make X::TypeCheck::Assignment message better
If we're assigning to an '@' or '%' sigiled container, mention that the
problem is with an element of the container. Potentially addresses #5110
so now something like `my Map %m = "abc", "def";` says `Type check
failed in assignment to an element of %m; expected Map but got Str
("def")` instead of just `Type check failed in assignment to %m;
expected Map but got Str ("def")`.