Introduce Least / Most
Work in progress as it unearthed a number of issues.

Introduce the dynamic variables $*LEAST and $*MOST as identifiers
for the smallest and largest possible value (like -Inf and Inf, but
without any typing).

In 6.c/d these will be initialized to -Inf and Inf respectively.

In 6.e, two new constants are made available:  Least and Most.  These
are just instances of Any that can be used for dispatching.  In 6.e
$*LEAST is initialized to Least, and $*MOST is initialized to Most.

Furthermore, candidates for infix operators are added to properly
handle the Least and Most values, so that any comparison with Least
and/or Most will yield the correct Boolean or Order enum.