Add optional parameter `$to` to locate-node
With `$to` the method will find not any narrowest node for a position,
but the narrowest for the specified range.

This solves a problem when we need to locate a node by position but that
position also matches one of its children. In this case the child is
chosen as being narrower that the parent. A typical example would be
`RakuAST::Call::Name` for which its `RakuAST::Name` child starts at the
same position.