Add Any.decont and decont() sub
Over the past weeks it has become clear that <> as a way to decont,
is very hard to understand for beginners, and very hard to make findable
in the documentation (because you will mostly find it under postcircumfix
handling).

This introduces an explicit .decont method on objects, and a decont()
sub that will call the .decont method on the given object.  It is set
up as a multi, so that classes can implement their own deconting logic.