Implement "use experimental :rakuast"
This pretty hacky approach fetches the RakuAST type object at the
end of the core.c compilation, and stashes that with bindcurhllsym.
It then replaces the CORE::<RakuAST> entry with Nil, causing all
lookups for RakuAST classes to fail from thereon.

It also lexically enables RakuAST again in the 6.e Formatter class
to keep that working.

And it adds an entry in lib/experimental that basically makes the
original RakuAST type object visible again.