Make sure that most deparsing features can be customized
Instead of having a method return an absolute string, return an
attribute of the Deparse object, that is initialized to the same
value.

No functional changes, just more flexibility towards the future.

Of note, added different handling RakuAST::Var::Compiler::File|Line
handling depending whether called as an instance or an object method:
the instance method will return the actual value, the class method
will return the code representation (aka $?FILE|$?LINE).  Not sure
whether this would be the right way to do it, but at least the
distinction is now made.