Correct line number for attribute accessor errors
Make it the attribute line.

NQP currently doesn't support `#line` for nqp-ops. To get around this
the line directive is set for `add_method` call. Also the call code is
is unrolled into a single line in order to prevent NQP from producing
wrong line numbers when (or if) the issue with ops gets fixed.