Avoid using a class as a role
Statement modified `for` used to use statement `for` method to produce
`QAST` implementation. Normally that would be delegated to a role, but
until we have them for RakuAST compiler let's use a parent class.

Prevents an attempt to set node's origin on an uninstantiated class and
fixes t/12-rakuast/statement-mods.t