Let the build be parallelizable
It's been done for NQP long ago, now it is possible for Rakudo. Not a
big gain to be expected as the primary time-consumers are CORE setting
targets and they cannot be done in parallel. But 15% speedup has been
observed on my setup.

Not well-tested for JVM backend which is still is not compatible with
RakuAST. But some preliminary steps has been done for it too.