Turn spectest operations into parallelizable targets
So that `make -j spectest` in a clean build directory would proceed as
expected. This also includes suppressing output from `git` so it doesn't
mess up with the rest of Makefile output. There was a problem about this approach,
as stated in [@niner
comment](https://github.com/rakudo/rakudo/pull/5142#discussion_r1056500111),
where git could stuck on a bad network. To make this situation
manageable git commands are now invoked through the monitoring script.
It would produce some heartbeat output every 30 seconds to entertain an
observer but would not allow git to remain silent for longer than 3 mins
or not finishing in 15 minutes.