Add basic RakuAST::Label tests
This is just for the object itself, tests where labels are actually
used inside code will need to be added later.  This appears to be
impossible atm, as the RakuAST::Label class appears to be missing a
IMPL-TO-QAST method implementation.

Also, it is unclear how RakuAST::Label should be used when building
ad-hoc structures.  Should it be a member of a StatementList, or an
attribute of a StatementList?