RakuAST: Hopefully finally fix BEGIN time execution hangs for good
The root of the problem is that we create QAST nodes much later than the old
compiler frontend. The old compiler could create a stub for code objects and
in the same process the QAST::Block object that generates the cuid. This cuid
is important when tracking clones of the code object. When a code object gets
compiled dynamically (i.e. at BEGIN time) we need to replace its $!do and the
$!do of all clones.

Previously we tried to get around this problem by ensuring that code
objects are linked before they may get cloned. However, it's become
clear that this approach just cannot work as e.g. a trait may make a
clone of the code object.

This new approach attacks the root: the need for creating the
QAST::Block to obtain the cuid. The block is not actually needed for
that. Instead we can generate a cuid, store it in the AST node and set
it on the QAST::Block when we create that.