RakuAST: fix crash with embedded comment at the start of the comp unit
lang_setup is run only after we parsed a possible language version
declaration which in turn may be preceeded by a comment. This comment
may need the literals builder, hence the crash.

Fix by setting up the literals builder in comp_unit_stage0 instead.
Since it needs the resolver, but the resolver needs the language
version, set the resolver on the literals builder once we have it.