Make IMPL-UNWRAP-LIST handle non-Listy things
By dropping them into a nqp::list.  This should make manual specification
of RakuAST nodes a bit more user-friendly, you wouldn't need to wrap them
into a [foo] or (foo,) if you only have a single thing to add

Please revert if this is considered to be a bad idea.