Simplify Supply / whenever processing order
Previously whenevers tapped their connected supply instantly. To prevent
deadlocks caused by loops back to the supply block the awaiter was
overwritten to take a continuation whenever a lock was awaited, snipping
off the stack up to the latest whenever. Those continuations were then put
in a list and run one after the other. This logic can work around all sorts
of locking scenarios, but is difficult to predict and can still manage to
deadlock in some cases. In addition this changed awaiter also triggered for
locks not part of the supply / whenever setup. So external unrelated locks
could affect the order of how the code is executed. That made the code even
more difficult to predict.

This patch removes all of that functionality. Instead whenevers now never
do their tap instantly, but always do it after the supply (or
whenever) code block containing the whenever has finished. That way the
supply block lock is released before the tapping happens that could
potentially also want the supplies lock, circumventing a lot of deadlock
potential.

As the new behavior does not at all try to solve deadlocks by reordering
execution order it is possible there are situations where the previous lock
solver would have resolved a situation, where the new logic deadlocks. I do
not know if there are any such practical cases.