Adapt heredoc tests
The strings should be specified *with* any leading spaces, however
the current implementation of RakuAST::Heredoc doesn't currently
handle those correctly.

Also, it's probably a bad idea to check the result of heredocs using
heredocs, so convert the tests to strings specified *without* heredocs.

Remove one test that was basically doing the same as another test.