Introduce the IGNORE phaser
This phaser will allow you to have the *runtime* completely ignore
a statement or a block.  Any code inside the IGNORE phaser **will**
be checked for syntax and legality, but will **not** cause any
bytecode to be generated.

This allows a developer to de-activate code (semi-)permanently,
while ensuring that the code inside this phaser remains in sync
with the rest of the code (which would not be the case if the
de-activated code would live on as comments).