RakuAST: fix bizarre miscompilations of role bodies
When roles are applied to classes within the same comp unit as they are
declared, they need to be fully compiled as their bodies are run when
the class is composed. However we only ran resolving, but did not apply
check time effects, so we e.g. did not thunk for loop expressions.

It looks like check time effects should just not modify the AST, i.e. not
try to thunk anything. That's better done as begin time effects as those
are guaranteed to run only once. Not sure if this will hold, but for now
it gets us further in the tests.