RakuAST: fix default type of signature declarations to Any
The default value for variables is Any, not Mu. However, we must not set
a non-Mu type on array or hash variables as we'd e.g. create an object
hash instead of a normal one. So instead of trying to figure out a
default type, just don't set any type if we don't have one, but instead
create containers for all variables unconditionally.

I don't think this is entirely correct, but with my existing confusion
about type, of and default, it's the best I can do and it unlocks a few
more tests.